written logic. need db and logger

+++ resources/entities/ohlcv.py
@@ -6,13 +6,14 @@ class OHLCV_candles:
 		self.C = list()
 		self.volume = list()
 
-	def add_new_OHLCV_candle(self, O, H, L, C, volume):
+	def add_new_OHLCV_candle(self, O, H, L, C, volume, delete_first_candle=True):
 		self.O.append(float(O))
 		self.H.append(float(H))
 		self.L.append(float(L))
 		self.C.append(float(C))
 		self.volume.append(float(volume))
-		self._delete_first_candle()
+		if delete_first_candle:
+			self._delete_first_candle()
 
 	def _delete_first_candle(self):
 		for lst in [self.O, self.H, self.L, self.C, self.volume]:
